feat(codegen): schema-first contract pipeline + all 8 engines generated
Introduce an all-Go schema-first codegen pipeline (cmd/schemagen, cmd/modelgen,
cmd/contractstrip; shared descriptor in cmd/internal/codegen) that reflects each
component's contract surface into a self-contained contract.schema.json,
regenerates the Go interface + models into contract.gen.go, and strips the
hand-written surface. Built on github.com/google/jsonschema-go (no Node).

All 8 engines now run on generated contracts (review, handoff, intervention,
settlement, billing, operationestimation, autoscaler, estimation), fully
encapsulated: each redefines its domain types as its own defs (zero projectstate
dependency); managers convert at the boundary. Enum String()/valid() methods
moved to free functions (behavior.go). Generator handles structs, enums
(type+const block), named scalars, external-domain inlining, map[string]T, and
time.Time / time.Duration / uuid.UUID. Pipeline is idempotent; the
TestMethodLayering conformance gate stays green.

+++ server/cmd/server/construction_adapters.go
@@ -377,7 +377,7 @@ func (a interventionAdapter) DecideOnVariance(
 		Kind:       interventionVarianceKind(v.Kind),
 		// AttemptCount drives Tiered retry-budget exhaustion (the Manager threads its
 		// supervision-loop attempt counter into the mirror's AttemptCount field).
-		AttemptCount: v.AttemptCount,
+		AttemptCount: int64(v.AttemptCount),
 		// Composition-supplied regime (Tiered{RetryBudget:2} default). Without a
 		// registered Mode the Engine rejects every variance with "unknown policy mode",
 		// so the composition root guarantees a non-Unknown mode is set.
